+++ scripts/inkscape.sh
@@ -2,5 +2,10 @@
 
 echo_info "Installing inkscape..."
 
+if command_exists inkscape; then
+    echo_success "Inkscape already installed!"
+    return
+fi
+
 # Install
 "${SUDO_CMD}" apt-get install -y inkscape
