+++ scripts/obsidian.sh
@@ -2,6 +2,11 @@
 
 echo_info "Installing Obsidian..."
 
+if command_exists obsidian; then
+    echo_success "Obsidian already installed!"
+    return
+fi
+
 wget https://github.com/obsidianmd/obsidian-releases/releases/download/v1.8.10/obsidian_1.8.10_amd64.deb
 "${SUDO_CMD}" apt-get install ./obsidian_1.8.10_amd64.deb
