+++ scripts/swaync.sh
@@ -2,4 +2,9 @@
 
 echo_info "Installing Notifications..."
 
+if command_exists swaync; then
+    echo_success "Swaync already installed!"
+    return
+fi
+
 "${SUDO_CMD}" apt-get install -y sway-notification-center libnotify-bin
