+++ scripts/vscode.sh
@@ -6,6 +6,11 @@ source $SCRIPT_DIR/check_env.sh
 
 echo_info "Installing VS Code..."
 
+if command_exists code; then
+    echo_success "VS Code already installed!"
+    return
+fi
+
 wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > packages.microsoft.gpg
 "${SUDO_CMD}" install -D -o root -g root -m 644 packages.microsoft.gpg /etc/apt/keyrings/packages.microsoft.gpg
 echo "deb [arch=amd64,arm64,armhf signed-by=/etc/apt/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/code stable main" | "${SUDO_CMD}" tee /etc/apt/sources.list.d/vscode.list > /dev/null
